backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

SQLAlchemy>=2.0.40
alembic>=1.14.0

httpx==0.26.0
python-dotenv==1.0.1
pydantic-settings==2.1.0

PyJWT==2.8.0
python-multipart==0.0.6
feedparser==6.0.11